chore(refactor): remove element after current def into separate function
