+++ lib/zeebe/VariableResolver.js
@@ -1,6 +1,9 @@
 import { getProcessVariables } from '@bpmn-io/extract-process-variables/zeebe';
 import { BaseVariableResolver } from '../base/VariableResolver';
-import { parseVariables } from './util/feelUtility';
+import {
+  parseVariables,
+  getElementNamesToRemove
+} from './util/feelUtility';
 import {
   getBusinessObject,
   is
@@ -36,42 +39,7 @@ export default class ZeebeVariableResolver extends BaseVariableResolver {
       return variables;
     }
 
-    const namesToFilter = [];
-
-    // Input: remove all inputs defined after the current input definition
-    if (is(moddleElement, 'zeebe:Input')) {
-      const allInputs = inputOutput.inputParameters;
-
-      const inputsToFilter =
-        allInputs
-          .slice(allInputs.indexOf(moddleElement))
-          .map(o => o.target);
-
-      namesToFilter.push(...inputsToFilter);
-    }
-
-    const allOutputs = inputOutput.outputParameters;
-
-    // Output: remove all outputs defined after the current output definition
-    if (is(moddleElement, 'zeebe:Output')) {
-
-      // Get all output mappings defined after the current element, including own name
-      const outputsToFilter = allOutputs
-        .slice(allOutputs.indexOf(moddleElement))
-        .map(o => o.target);
-
-      namesToFilter.push(...outputsToFilter);
-    }
-
-    // Input or general property: remove all outputs
-    else if (allOutputs) {
-
-      // Input or execution-related element, remove all outputs
-      const outputsToFilter = allOutputs
-        .map(o => o.target);
-
-      namesToFilter.push(...outputsToFilter);
-    }
+    const namesToFilter = getElementNamesToRemove(moddleElement, inputOutput);
 
     return variables.filter(v => {
