+++ lib/zeebe/util/feelUtility.js
@@ -6,6 +6,7 @@ import {
 import { EntriesContext } from './VariableContext';
 import { getExtensionElementsList } from '../../base/util/ExtensionElementsUtil';
 import { getParents } from '../../base/util/scopeUtil';
+import { is } from 'bpmn-js/lib/util/ModelUtil';
 
 export function parseVariables(variables) {
 
@@ -374,4 +375,48 @@ function filterForScope(context, variable) {
   }
 
   return scopedResults;
+}
+
+/**
+ * Remove input/output element name after current definition
+ */
+export function getElementNamesToRemove(moddleElement, inputOutput) {
+  const namesToFilter = [];
+
+  // Input: remove all inputs defined after the current input definition
+  if (is(moddleElement, 'zeebe:Input')) {
+    const allInputs = inputOutput.inputParameters;
+
+    const inputsToFilter =
+        allInputs
+          .slice(allInputs.indexOf(moddleElement))
+          .map(o => o.target);
+
+    namesToFilter.push(...inputsToFilter);
+  }
+
+  const allOutputs = inputOutput.outputParameters;
+
+  // Output: remove all outputs defined after the current output definition
+  if (is(moddleElement, 'zeebe:Output')) {
+
+    // Get all output mappings defined after the current element, including own name
+    const outputsToFilter = allOutputs
+      .slice(allOutputs.indexOf(moddleElement))
+      .map(o => o.target);
+
+    namesToFilter.push(...outputsToFilter);
+  }
+
+  // Input or general property: remove all outputs
+  else if (allOutputs) {
+
+    // Input or execution-related element, remove all outputs
+    const outputsToFilter = allOutputs
+      .map(o => o.target);
+
+    namesToFilter.push(...outputsToFilter);
+  }
+
+  return namesToFilter;
 }
